Sending text and validation with focus change on Desktop

+++ common/src/main/kotlin/communication/DTO.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 
 
 @Serializable
-class DTOText(val id: Long, val text: String, val label: String)
+class DTOText(val id: Int, val text: String, val label: String)
 
 @Serializable
 class DTOValidation(val onRightTrack: Boolean = true, val isValid: Boolean = true,
